+++ shuup/admin/forms/_base.py
@@ -8,7 +8,7 @@
 
 from filer.fields.file import AdminFileWidget
 
-from shuup.admin.forms.widgets import MediaChoiceWidget
+from shuup.admin.forms.widgets import FileDnDUploaderWidget
 from shuup.utils.multilanguage_model_form import MultiLanguageModelForm
 
 
@@ -18,4 +18,5 @@ class ShuupAdminForm(MultiLanguageModelForm):
         super(ShuupAdminForm, self).__init__(**kwargs)
         for field in self.fields:
             if issubclass(self.fields[field].widget.__class__, AdminFileWidget):
-                self.fields[field].widget = MediaChoiceWidget(clearable=True)
+                self.fields[field].widget = FileDnDUploaderWidget(
+                    upload_path="/default", kind="images", clearable=True)