+++ doc/changelog.rst
@@ -21,14 +21,19 @@ Localization
 Admin
 ~~~~~
 
+- Add dropzone widget for shop, category, service provider
+  and service image fields
+- Add option to clear dropzone selection
 - Add option to install sample data in Wizard
-
 
 Addons
 ~~~~~~
 
 Front
 ~~~~~
+
+- Make carousel slide available by default
+- Add dropzone widget for carousel slide images
 
 Xtheme
 ~~~~~~